Updated plugin and specs
